requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
